+++ src/user/controller/user.controller.js
@@ -4,6 +4,18 @@ import { UserSignupRequestDto } from "../dto/user.dto.js";
 import { UserLoginRequestDto } from "../dto/user.dto.js";
 import { verifyJwt } from "../../jwt.config.js";
 
+// 로그인 요청 
+
+export const userLogin = async(req, res, next) => {
+    try{
+        const dto = new UserLoginRequestDto(req.body);
+        const result = await UserService.userLogin(dto);
+        res.status(StatusCodes.OK).success(result);
+    }catch(err) {
+        next(err);
+    }
+}
+
 // 사용자(계정) 추가 
 export const addUser = async(req, res, next) => {
     try{
@@ -24,18 +36,6 @@ export const addUser = async(req, res, next) => {
         // 3. 성공 응답 반환
         res.status(StatusCodes.CREATED).success(result);
     } catch(err) {
-        next(err);
-    }
-}
-
-// 로그인 요청 
-
-export const userLogin = async(req, res, next) => {
-    try{
-        const dto = new UserLoginRequestDto(req.body);
-        const result = await UserService.userLogin(dto);
-        res.status(StatusCodes.OK).success(result);
-    }catch(err) {
         next(err);
     }
 }
