+++ src/user/dto/user.dto.js
@@ -1,3 +1,11 @@
+// 000으로 시작하기 버튼 클릭 시 
+export class UserLoginRequestDto {
+    constructor({provider, accessToken}) {
+        this.provider = provider;
+        this.accessToken = accessToken;
+    }
+}
+
 // 회원가입 요청
 export class UserSignupRequestDto {
     constructor({nickname,provider, oauth_id, categories, agreements}) {
@@ -7,11 +15,4 @@ export class UserSignupRequestDto {
         this.categories = categories; // 배열
         this.agreements = agreements; // 배열
     }
-}
-
-export class UserLoginRequestDto {
-  constructor({provider, accessToken}) {
-    this.provider = provider;
-    this.accessToken = accessToken;
-  }
 }
